General: Use java.net.URI instead of URL for string validation
